+++ src/extension.ts
@@ -1,25 +1,25 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
-import * as vscode from 'vscode'
+import * as vscode from 'vscode';
 
 // Import the Configs, Controllers, and Providers
 import {
   EXTENSION_DISPLAY_NAME,
   EXTENSION_ID,
   ExtensionConfig,
-} from './app/configs'
+} from './app/configs';
 import {
   ExampleController,
   FeedbackController,
   ListFilesController,
-} from './app/controllers'
+} from './app/controllers';
 import {
   ChatProvider,
   ColorProvider,
   FeedbackProvider,
   ListFilesProvider,
-} from './app/providers'
-import { OpenAIService } from './app/services'
+} from './app/providers';
+import { OpenAIService } from './app/services';
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
@@ -29,11 +29,11 @@ export function activate(context: vscode.ExtensionContext) {
     | vscode.Uri
     | vscode.TextDocument
     | vscode.WorkspaceFolder
-    | undefined
+    | undefined;
 
   // Get the resource for the workspace
   if (vscode.workspace.workspaceFolders) {
-    resource = vscode.workspace.workspaceFolders[0]
+    resource = vscode.workspace.workspaceFolders[0];
   }
 
   // -----------------------------------------------------------------
@@ -41,31 +41,31 @@ export function activate(context: vscode.ExtensionContext) {
   // -----------------------------------------------------------------
 
   // Get the previous version of the extension
-  const previousVersion = context.globalState.get('version')
+  const previousVersion = context.globalState.get('version');
   // Get the current version of the extension
-  const currentVersion = context.extension.packageJSON.version
+  const currentVersion = context.extension.packageJSON.version;
 
-  console.log('Previous version:', previousVersion)
-  console.log('Extension version:', currentVersion)
+  console.log('Previous version:', previousVersion);
+  console.log('Extension version:', currentVersion);
 
   // Check if the extension is running for the first time
   if (!previousVersion) {
     vscode.window.showInformationMessage(
       `Welcome to ${EXTENSION_DISPLAY_NAME}!`,
-    )
+    );
 
     // Update the version in the global state
-    context.globalState.update('version', currentVersion)
+    context.globalState.update('version', currentVersion);
   }
 
   // Check if the extension has been updated
   if (previousVersion !== currentVersion) {
     vscode.window.showInformationMessage(
       `Looks like ${EXTENSION_DISPLAY_NAME} has been updated to version ${currentVersion}!`,
-    )
+    );
 
     // Update the version in the global state
-    context.globalState.update('version', currentVersion)
+    context.globalState.update('version', currentVersion);
   }
 
   // -----------------------------------------------------------------
@@ -75,7 +75,7 @@ export function activate(context: vscode.ExtensionContext) {
   // Get the configuration for the extension
   const config = new ExtensionConfig(
     vscode.workspace.getConfiguration(EXTENSION_ID, resource),
-  )
+  );
 
   // #region Example
 
@@ -84,7 +84,7 @@ export function activate(context: vscode.ExtensionContext) {
   // -----------------------------------------------------------------
 
   // Create a new ExampleController
-  const exampleController = new ExampleController(config)
+  const exampleController = new ExampleController(config);
 
   // The command has been defined in the package.json file
   // Now provide the implementation of the command with registerCommand
@@ -95,9 +95,9 @@ export function activate(context: vscode.ExtensionContext) {
       // The code you place here will be executed every time your command is executed
 
       // Display a message box to the user
-      exampleController.helloWorld()
+      exampleController.helloWorld();
     },
-  )
+  );
 
   // Register the command to get the files in a folder
   // The commandId parameter must match the command field in package.json
@@ -105,18 +105,18 @@ export function activate(context: vscode.ExtensionContext) {
   const disposableGetFilesInFolder = vscode.commands.registerCommand(
     `${EXTENSION_ID}.getFilesInFolder`,
     (args) => exampleController.getFilesInFolder(args),
-  )
+  );
 
   const disposableConvertToTS = vscode.commands.registerCommand(
     `${EXTENSION_ID}.convertToTS`,
     () => exampleController.convertToTS(),
-  )
+  );
 
   context.subscriptions.push(
     disposableHelloWorld,
     disposableGetFilesInFolder,
     disposableConvertToTS,
-  )
+  );
 
   // #endregion Example
 
@@ -127,21 +127,21 @@ export function activate(context: vscode.ExtensionContext) {
   // -----------------------------------------------------------------
 
   // Create a new ListFilesController
-  const listFilesController = new ListFilesController(config)
+  const listFilesController = new ListFilesController(config);
 
   const disposableOpenFile = vscode.commands.registerCommand(
     `${EXTENSION_ID}.listFiles.openFile`,
     (uri) => listFilesController.openFile(uri),
-  )
+  );
 
-  context.subscriptions.push(disposableOpenFile)
+  context.subscriptions.push(disposableOpenFile);
 
   // -----------------------------------------------------------------
   // Register ListFilesProvider and list commands
   // -----------------------------------------------------------------
 
   // Create a new ListFilesProvider
-  const listFilesProvider = new ListFilesProvider(listFilesController)
+  const listFilesProvider = new ListFilesProvider(listFilesController);
 
   // Register the list provider
   const listFilesTreeView = vscode.window.createTreeView(
@@ -150,34 +150,34 @@ export function activate(context: vscode.ExtensionContext) {
       treeDataProvider: listFilesProvider,
       showCollapseAll: true,
     },
-  )
+  );
 
   const disposableRefreshList = vscode.commands.registerCommand(
     `${EXTENSION_ID}.listFiles.refreshList`,
     () => listFilesProvider.refresh(),
-  )
+  );
 
-  context.subscriptions.push(listFilesTreeView, disposableRefreshList)
+  context.subscriptions.push(listFilesTreeView, disposableRefreshList);
 
   // -----------------------------------------------------------------
   // Register ListFilesProvider and ListMethodsProvider events
   // -----------------------------------------------------------------
 
   vscode.workspace.onDidChangeTextDocument(() => {
-    listFilesProvider.refresh()
-  })
+    listFilesProvider.refresh();
+  });
   vscode.workspace.onDidCreateFiles(() => {
-    listFilesProvider.refresh()
-  })
+    listFilesProvider.refresh();
+  });
   vscode.workspace.onDidDeleteFiles(() => {
-    listFilesProvider.refresh()
-  })
+    listFilesProvider.refresh();
+  });
   vscode.workspace.onDidRenameFiles(() => {
-    listFilesProvider.refresh()
-  })
+    listFilesProvider.refresh();
+  });
   vscode.workspace.onDidSaveTextDocument(() => {
-    listFilesProvider.refresh()
-  })
+    listFilesProvider.refresh();
+  });
 
   // #endregion List Files
 
@@ -188,7 +188,7 @@ export function activate(context: vscode.ExtensionContext) {
   // -----------------------------------------------------------------
 
   // Create a new FeedbackProvider
-  const feedbackProvider = new FeedbackProvider(new FeedbackController())
+  const feedbackProvider = new FeedbackProvider(new FeedbackController());
 
   // Register the feedback provider
   const feedbackTreeView = vscode.window.createTreeView(
@@ -196,33 +196,33 @@ export function activate(context: vscode.ExtensionContext) {
     {
       treeDataProvider: feedbackProvider,
     },
-  )
+  );
 
   // Register the commands
   const disposableAboutUs = vscode.commands.registerCommand(
     `${EXTENSION_ID}.feedback.aboutUs`,
     () => feedbackProvider.controller.aboutUs(),
-  )
+  );
   const disposableDocumentation = vscode.commands.registerCommand(
     `${EXTENSION_ID}.feedback.documentation`,
     () => feedbackProvider.controller.documentation(),
-  )
+  );
   const disposableReportIssues = vscode.commands.registerCommand(
     `${EXTENSION_ID}.feedback.reportIssues`,
     () => feedbackProvider.controller.reportIssues(),
-  )
+  );
   const disposableRateUs = vscode.commands.registerCommand(
     `${EXTENSION_ID}.feedback.rateUs`,
     () => feedbackProvider.controller.rateUs(),
-  )
+  );
   const disposableFollowUs = vscode.commands.registerCommand(
     `${EXTENSION_ID}.feedback.followUs`,
     () => feedbackProvider.controller.followUs(),
-  )
+  );
   const disposableSupportUs = vscode.commands.registerCommand(
     `${EXTENSION_ID}.feedback.supportUs`,
     () => feedbackProvider.controller.supportUs(),
-  )
+  );
 
   context.subscriptions.push(
     feedbackTreeView,
@@ -232,7 +232,7 @@ export function activate(context: vscode.ExtensionContext) {
     disposableRateUs,
     disposableFollowUs,
     disposableSupportUs,
-  )
+  );
 
   // #endregion Feedback
 
@@ -243,16 +243,16 @@ export function activate(context: vscode.ExtensionContext) {
   // -----------------------------------------------------------------
 
   // Create a new ChatProvider
-  const chatProvider = new ChatProvider(context.extensionUri)
-  chatProvider.setService(new OpenAIService(config))
+  const chatProvider = new ChatProvider(context.extensionUri);
+  chatProvider.setService(new OpenAIService(config));
 
   // Register the ChatProvider
   const chatWebviewProvider = vscode.window.registerWebviewViewProvider(
     ChatProvider.viewType,
     chatProvider,
-  )
+  );
 
-  context.subscriptions.push(chatWebviewProvider)
+  context.subscriptions.push(chatWebviewProvider);
 
   // #endregion Chat
 
@@ -263,33 +263,33 @@ export function activate(context: vscode.ExtensionContext) {
   // -----------------------------------------------------------------
 
   // Create a new ColorProvider
-  const colorProvider = new ColorProvider(context.extensionUri)
+  const colorProvider = new ColorProvider(context.extensionUri);
 
   // Register the ColorProvider
   const colorWebviewProvider = vscode.window.registerWebviewViewProvider(
     ColorProvider.viewType,
     colorProvider,
-  )
+  );
 
   const disposableAddColor = vscode.commands.registerCommand(
     `${EXTENSION_ID}.colors.addColor`,
     () => {
-      colorProvider.addColor()
+      colorProvider.addColor();
     },
-  )
+  );
 
   const disposableClearColors = vscode.commands.registerCommand(
     `${EXTENSION_ID}.colors.clearColors`,
     () => {
-      colorProvider.clearColors()
+      colorProvider.clearColors();
     },
-  )
+  );
 
   context.subscriptions.push(
     colorWebviewProvider,
     disposableAddColor,
     disposableClearColors,
-  )
+  );
 
   // #endregion Color
 }